requirements.txt
azure-functions
azure-functions-durable
pyodbc
simplejson